Move tests from PR #181 under models and views test dirs accordingly
